refactor(agent): expose capabilities from subpath
