+++ packages/agent/src/types.ts
@@ -266,7 +266,6 @@ type AgentSettingsBase<
   model?: AgentModelResolver<TRuntimeConfig>
   provider?: AgentModelProvider
   runtime?: AgentRuntimeBinding
-  tools?: never
   workspace?: WorkspaceAgentWorkspaceConfig
   [key: string]: unknown
 }
